backend/requirements.txt
fastapi==1.0.0
uvicorn==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
google-generativeai==0.3.1
python-multipart==0.0.6
pypdf==3.16.2
python-docx==0.8.11
python-dotenv==1.0.0
